requirements.txt
# Framework web y servidor ASGI
fastapi==0.135.0
uvicorn[standard]==0.41.0
pydantic==2.12.5
aiofiles==25.1.0

# Motor de plantillas
pybars3==0.9.7

# Cliente HTTP (usado por varios componentes)
requests==2.32.5

# Variables de entorno
python-dotenv==1.2.2
